Food_Line_bot/darknet/requirements.txt
asgiref==3.3.1
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
blinker==1.4
boto==2.49.0
certifi==2018.1.18
cffi==1.14.5
chardet==3.0.4
click==6.7

colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.6
defusedxml==0.6.0
diff-match-patch==20200713
distro==1.5.0
Django==3.1.6
django-allauth==0.44.0
django-grappelli==2.14.3
django-import-export==2.5.0
django-scheduler==0.9.3
et-xmlfile==1.0.1
Flask==1.1.2
future==0.18.2
google-compute-engine==2.8.13
httplib2==0.9.2
hyperlink==17.3.1
icalendar==4.0.7
idna==2.6
incremental==16.10.1
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.3
joblib==1.0.1
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
liffpy==0.0.1a3
line-bot-sdk==1.18.0
MarkupPy==1.14
MarkupSafe==1.0
netifaces==0.10.4
ngrok==0.0.1
numpy==1.19.5
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.6
pandas==1.1.5
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.20
pycrypto==2.6.1
pyimgur==0.6.0
PyJWT==2.0.1
pyOpenSSL==17.5.0
pyserial==3.4
python-dateutil==2.8.1
python-debian==0.1.32
python3-openid==3.2.0
pytz==2021.1
pyxdg==0.25
Pillow==8.1.2
psycopg2

requests==2.18.4
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
scikit-learn==0.24.1
scipy==1.5.4
SecretStorage==2.3.1
service-identity==16.0.0
six==1.11.0
sklearn==0.0
sqlparse==0.4.1
# systemd-python==234
tablib==3.0.0
threadpoolctl==2.1.0
Twisted==17.9.0
urllib3==1.22
Werkzeug==1.0.1
xlrd==2.0.1
xlwt==1.3.0
